Completed project
Project Euler problem set (multiples of 3 or 5). This marks the beginning of my journey with algorithms. You'll note that my first solution was a basic for loop and if statement. The pdf that contained the solution, however, taught me a better way using a proof of Karl Gauss' theory summing consecutive integers.